arm64: dts: qcom: sm8150: Add msm bus apps rsc device
Since the Resource Power Manager Hardened (RPMH) driver is no longer
using the mailbox framework. Create a new device as the child node
of the RPMH driver to help fetch handles for each Resource State
Coordinator(RSC). This is required to interface with the RPMH driver
in order to communicate with a set of Qualcomm specific ASICs
controlling bus clock resources.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>