dm: default-key: Adapt legacy disk format for new set of arguments
User now have option to choose between legacy and new version of
on disk data format. So arguments passed in case of legacy format
will not match with the count check in default key driver. So
adapt the legacy support for new set of arguments.

Test:
1) Set fileencryption=ice.
2) Compile the code.
3) Flash the generated images.
4) Make sure the device boot normally.
5) Make sure there is no error output like the following
in the log:
unspported encryption mode.

Change-Id: Ie6f9f683c048a2a1c64d46716433bcff359dc3bf
Signed-off-by: mxing <quic_mxing@quicinc.com>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>